Rename and move card expansion
