+++ angular-frontend/src/app/features/issuer/issuer.module.ts
@@ -4,8 +4,6 @@ import { SharedModule } from '@shared/shared.module';
 
 import { IssuerRoutingModule } from './issuer-routing.module';
 import { CredentialsPageComponent } from './pages/credentials-page/credentials-page.component';
-import { IssuancePageComponent } from './pages/issuance-page/issuance-page.component';
-import { MatCardExpansionComponent } from './shared/components/mat-card-expansion/mat-card-expansion.component';
 import { PageHeaderComponent } from './shared/components/page-header/page-header.component';
 import { PageHeaderSummaryDirective } from './shared/components/page-header/page-header-summary.directive';
 
@@ -16,8 +14,6 @@ import { PageHeaderSummaryDirective } from './shared/components/page-header/page
   ],
   declarations: [
     CredentialsPageComponent,
-    IssuancePageComponent,
-    MatCardExpansionComponent,
     PageHeaderComponent,
     PageHeaderSummaryDirective
   ]
