+++ angular-frontend/src/app/features/issuer/pages/credentials-page/credentials-page.component.html
@@ -52,7 +52,7 @@
               </mat-list-item>
             </mat-list>
 
-            <app-mat-card-expansion [immunizationRecord]="record"></app-mat-card-expansion>
+            <app-card-action-expansion [immunizationRecord]="record"></app-card-action-expansion>
           </mat-card-content>
         </app-card-alert>
