--- angular-frontend/src/app/features/issuer/pages/issuance-page/issuance-page.component.html
@@ -1,2 +1,0 @@
-<button mat-flat-button
-        routerLink="/issuer/issuance">Back</button>
