--- angular-frontend/src/app/features/issuer/pages/issuance-page/issuance-page.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { IssuancePageComponent } from './issuance-page.component';
-
-describe('IssuancePageComponent', () => {
-  let component: IssuancePageComponent;
-  let fixture: ComponentFixture<IssuancePageComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [IssuancePageComponent]
-    }).compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(IssuancePageComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
