--- angular-frontend/src/app/features/issuer/pages/issuance-page/issuance-page.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-issuance-page',
-  templateUrl: './issuance-page.component.html',
-  styleUrls: ['./issuance-page.component.scss']
-})
-export class IssuancePageComponent implements OnInit {
-  public constructor( ) { }
-
-  public ngOnInit(): void { }
-}
