+++ angular-frontend/src/app/shared/components/card-action-expansion/card-action-expansion.component.html
@@ -10,6 +10,17 @@
   </button>
 </mat-card-actions>
 
+<div class="expandable">
+  <button mat-button
+          class="adjust-margin"
+          (click)="expanded = !expanded">
+    {{ (expanded) ? 'Collapse' : 'Expand' }}
+    <mat-icon>
+      {{ (expanded) ? 'expand_less' : 'expand_more' }}
+    </mat-icon>
+  </button>
+</div>
+
 <ng-container *ngIf="expanded">
   <mat-list dense>
