+++ angular-frontend/src/app/shared/components/card-action-expansion/card-action-expansion.component.scss
