+++ angular-frontend/src/app/shared/components/card-action-expansion/card-action-expansion.component.ts
@@ -1,0 +1,17 @@
+import { Component, Input } from '@angular/core';
+
+import { ImmunizationRecord } from '../../../features/issuer/shared/models/immunization-record.model';
+
+@Component({
+  selector: 'app-card-action-expansion',
+  templateUrl: './card-action-expansion.component.html',
+  styleUrls: ['./card-action-expansion.component.scss']
+})
+export class CardActionExpansionComponent {
+  @Input() public immunizationRecord!: ImmunizationRecord;
+  public expanded: boolean;
+
+  public constructor() {
+    this.expanded = false;
+  }
+}
