+++ angular-frontend/src/app/shared/shared.module.ts
@@ -7,6 +7,7 @@ import { NgxMaterialModule } from './modules/ngx-material/ngx-material.module';
 import { FormatDatePipe } from './pipes/format-date.pipe';
 import { SafePipe } from './pipes/safe.pipe';
 import { CardAlertComponent } from './components/card-alert/card-alert.component';
+import { CardActionExpansionComponent } from './components/card-action-expansion/card-action-expansion.component';
 import { QRCodeComponent } from './components/qrcode/qrcode.component';
 
 @NgModule({
@@ -14,6 +15,7 @@ import { QRCodeComponent } from './components/qrcode/qrcode.component';
     FormatDatePipe,
     SafePipe,
     CardAlertComponent,
+    CardActionExpansionComponent,
     QRCodeComponent
   ],
   imports: [
@@ -29,6 +31,7 @@ import { QRCodeComponent } from './components/qrcode/qrcode.component';
     FormatDatePipe,
     SafePipe,
     CardAlertComponent,
+    CardActionExpansionComponent,
     QRCodeComponent
   ]
 })
